# gen/ogen/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package ogen

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [1]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"
			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'c': // Prefix: "comments/"
				if l := len("comments/"); len(elem) >= l && elem[0:l] == "comments/" {
					elem = elem[l:]
				} else {
					break
				}

				// Param: "joke_id"
				// Leaf parameter
				args[0] = elem
				elem = ""

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleCommentsByJokeIDGetRequest([1]string{
							args[0],
						}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET")
					}

					return
				}
			case 'l': // Prefix: "likes/"
				if l := len("likes/"); len(elem) >= l && elem[0:l] == "likes/" {
					elem = elem[l:]
				} else {
					break
				}

				// Param: "joke_id"
				// Leaf parameter
				args[0] = elem
				elem = ""

				if len(elem) == 0 {
					// Leaf node.
					switch r.Method {
					case "GET":
						s.handleLikesByJokeIDGetRequest([1]string{
							args[0],
						}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "GET")
					}

					return
				}
			case 'm': // Prefix: "me"
				if l := len("me"); len(elem) >= l && elem[0:l] == "me" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch r.Method {
					case "DELETE":
						s.handleMeDelRequest([0]string{}, elemIsEscaped, w, r)
					case "GET":
						s.handleMeGetRequest([0]string{}, elemIsEscaped, w, r)
					case "PUT":
						s.handleMePutRequest([0]string{}, elemIsEscaped, w, r)
					default:
						s.notAllowed(w, r, "DELETE,GET,PUT")
					}

					return
				}
				switch elem[0] {
				case '/': // Prefix: "/"
					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'c': // Prefix: "comments"
						if l := len("comments"); len(elem) >= l && elem[0:l] == "comments" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "DELETE":
								s.handleMeCommentsDelRequest([0]string{}, elemIsEscaped, w, r)
							case "POST":
								s.handleMeCommentsPostRequest([0]string{}, elemIsEscaped, w, r)
							case "PUT":
								s.handleMeCommentsPutRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, "DELETE,POST,PUT")
							}

							return
						}
					case 'j': // Prefix: "jokes"
						if l := len("jokes"); len(elem) >= l && elem[0:l] == "jokes" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "DELETE":
								s.handleMeJokesDelRequest([0]string{}, elemIsEscaped, w, r)
							case "GET":
								s.handleMeJokesGetRequest([0]string{}, elemIsEscaped, w, r)
							case "POST":
								s.handleMeJokesPostRequest([0]string{}, elemIsEscaped, w, r)
							case "PUT":
								s.handleMeJokesPutRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, "DELETE,GET,POST,PUT")
							}

							return
						}
					case 'l': // Prefix: "likes"
						if l := len("likes"); len(elem) >= l && elem[0:l] == "likes" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							// Leaf node.
							switch r.Method {
							case "DELETE":
								s.handleMeLikesDelRequest([0]string{}, elemIsEscaped, w, r)
							case "POST":
								s.handleMeLikesPostRequest([0]string{}, elemIsEscaped, w, r)
							default:
								s.notAllowed(w, r, "DELETE,POST")
							}

							return
						}
					}
				}
			case 's': // Prefix: "sign_"
				if l := len("sign_"); len(elem) >= l && elem[0:l] == "sign_" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'i': // Prefix: "in"
					if l := len("in"); len(elem) >= l && elem[0:l] == "in" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleSignInPostRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "POST")
						}

						return
					}
				case 'r': // Prefix: "refresh"
					if l := len("refresh"); len(elem) >= l && elem[0:l] == "refresh" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleSignRefreshPostRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "POST")
						}

						return
					}
				case 'u': // Prefix: "up"
					if l := len("up"); len(elem) >= l && elem[0:l] == "up" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						// Leaf node.
						switch r.Method {
						case "POST":
							s.handleSignUpPostRequest([0]string{}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, "POST")
						}

						return
					}
				}
			}
		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name        string
	summary     string
	operationID string
	pathPattern string
	count       int
	args        [1]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/"
			if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				break
			}
			switch elem[0] {
			case 'c': // Prefix: "comments/"
				if l := len("comments/"); len(elem) >= l && elem[0:l] == "comments/" {
					elem = elem[l:]
				} else {
					break
				}

				// Param: "joke_id"
				// Leaf parameter
				args[0] = elem
				elem = ""

				if len(elem) == 0 {
					switch method {
					case "GET":
						// Leaf: CommentsByJokeIDGet
						r.name = "CommentsByJokeIDGet"
						r.summary = "Returns comments of the joke"
						r.operationID = "CommentsByJokeIDGet"
						r.pathPattern = "/comments/{joke_id}"
						r.args = args
						r.count = 1
						return r, true
					default:
						return
					}
				}
			case 'l': // Prefix: "likes/"
				if l := len("likes/"); len(elem) >= l && elem[0:l] == "likes/" {
					elem = elem[l:]
				} else {
					break
				}

				// Param: "joke_id"
				// Leaf parameter
				args[0] = elem
				elem = ""

				if len(elem) == 0 {
					switch method {
					case "GET":
						// Leaf: LikesByJokeIDGet
						r.name = "LikesByJokeIDGet"
						r.summary = "Counts likes of the joke"
						r.operationID = "LikesByJokeIDGet"
						r.pathPattern = "/likes/{joke_id}"
						r.args = args
						r.count = 1
						return r, true
					default:
						return
					}
				}
			case 'm': // Prefix: "me"
				if l := len("me"); len(elem) >= l && elem[0:l] == "me" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					switch method {
					case "DELETE":
						r.name = "MeDel"
						r.summary = "Deletes current user profile"
						r.operationID = "MeDel"
						r.pathPattern = "/me"
						r.args = args
						r.count = 0
						return r, true
					case "GET":
						r.name = "MeGet"
						r.summary = "Returns current user profile"
						r.operationID = "MeGet"
						r.pathPattern = "/me"
						r.args = args
						r.count = 0
						return r, true
					case "PUT":
						r.name = "MePut"
						r.summary = "Updates current user profile"
						r.operationID = "MePut"
						r.pathPattern = "/me"
						r.args = args
						r.count = 0
						return r, true
					default:
						return
					}
				}
				switch elem[0] {
				case '/': // Prefix: "/"
					if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						break
					}
					switch elem[0] {
					case 'c': // Prefix: "comments"
						if l := len("comments"); len(elem) >= l && elem[0:l] == "comments" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							switch method {
							case "DELETE":
								// Leaf: MeCommentsDel
								r.name = "MeCommentsDel"
								r.summary = "Deletes a comment of the current user"
								r.operationID = "MeCommentsDel"
								r.pathPattern = "/me/comments"
								r.args = args
								r.count = 0
								return r, true
							case "POST":
								// Leaf: MeCommentsPost
								r.name = "MeCommentsPost"
								r.summary = "Creates a comment for the current user and the current joke"
								r.operationID = "MeCommentsPost"
								r.pathPattern = "/me/comments"
								r.args = args
								r.count = 0
								return r, true
							case "PUT":
								// Leaf: MeCommentsPut
								r.name = "MeCommentsPut"
								r.summary = "Updates a comment of the current user"
								r.operationID = "MeCommentsPut"
								r.pathPattern = "/me/comments"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}
					case 'j': // Prefix: "jokes"
						if l := len("jokes"); len(elem) >= l && elem[0:l] == "jokes" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							switch method {
							case "DELETE":
								// Leaf: MeJokesDel
								r.name = "MeJokesDel"
								r.summary = "Deletes joke for current user"
								r.operationID = "MeJokesDel"
								r.pathPattern = "/me/jokes"
								r.args = args
								r.count = 0
								return r, true
							case "GET":
								// Leaf: MeJokesGet
								r.name = "MeJokesGet"
								r.summary = "Returns jokes of the current user"
								r.operationID = "MeJokesGet"
								r.pathPattern = "/me/jokes"
								r.args = args
								r.count = 0
								return r, true
							case "POST":
								// Leaf: MeJokesPost
								r.name = "MeJokesPost"
								r.summary = "Creates a new joke for current user"
								r.operationID = "MeJokesPost"
								r.pathPattern = "/me/jokes"
								r.args = args
								r.count = 0
								return r, true
							case "PUT":
								// Leaf: MeJokesPut
								r.name = "MeJokesPut"
								r.summary = "Updates joke for current user"
								r.operationID = "MeJokesPut"
								r.pathPattern = "/me/jokes"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}
					case 'l': // Prefix: "likes"
						if l := len("likes"); len(elem) >= l && elem[0:l] == "likes" {
							elem = elem[l:]
						} else {
							break
						}

						if len(elem) == 0 {
							switch method {
							case "DELETE":
								// Leaf: MeLikesDel
								r.name = "MeLikesDel"
								r.summary = "Deletes a like of the current user"
								r.operationID = "MeLikesDel"
								r.pathPattern = "/me/likes"
								r.args = args
								r.count = 0
								return r, true
							case "POST":
								// Leaf: MeLikesPost
								r.name = "MeLikesPost"
								r.summary = "Creates a like for a joke for the current user"
								r.operationID = "MeLikesPost"
								r.pathPattern = "/me/likes"
								r.args = args
								r.count = 0
								return r, true
							default:
								return
							}
						}
					}
				}
			case 's': // Prefix: "sign_"
				if l := len("sign_"); len(elem) >= l && elem[0:l] == "sign_" {
					elem = elem[l:]
				} else {
					break
				}

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case 'i': // Prefix: "in"
					if l := len("in"); len(elem) >= l && elem[0:l] == "in" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch method {
						case "POST":
							// Leaf: SignInPost
							r.name = "SignInPost"
							r.summary = "Performs user authentication"
							r.operationID = "SignInPost"
							r.pathPattern = "/sign_in"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}
				case 'r': // Prefix: "refresh"
					if l := len("refresh"); len(elem) >= l && elem[0:l] == "refresh" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch method {
						case "POST":
							// Leaf: SignRefreshPost
							r.name = "SignRefreshPost"
							r.summary = "Exchanges a refresh token for an access token"
							r.operationID = "SignRefreshPost"
							r.pathPattern = "/sign_refresh"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}
				case 'u': // Prefix: "up"
					if l := len("up"); len(elem) >= l && elem[0:l] == "up" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch method {
						case "POST":
							// Leaf: SignUpPost
							r.name = "SignUpPost"
							r.summary = "Performs user registration"
							r.operationID = "SignUpPost"
							r.pathPattern = "/sign_up"
							r.args = args
							r.count = 0
							return r, true
						default:
							return
						}
					}
				}
			}
		}
	}
	return r, false
}