# gen/openapi/openapi.yaml
openapi: 3.0.0
info:
  version: 0.1.0
  title: Flugo API
  termsOfService: https://github.com/abc-valera/flugo-api-golang
  contact:
    email: zoocity14@gmail.com
    url: https://github.com/abc-valera/flugo-api-golang
  description: API for Flugo social network
servers:
  - url: http://localhost:3000
tags:
  - name: sign
    description: Everything about managing user sessions
  - name: users
    description: Everything about searching users
  - name: jokes
    description: Everything about searching jokes
  - name: comments
    description: Everything about searching comments
  - name: likes
    description: Everything about counting likes
  - name: me
    description: Everything about managing current user profile
externalDocs:
  description: Learn more about the project.
  url: https://github.com/abc-valera/flugo-api-golang
paths:
  /sign_up:
    post:
      tags:
        - sign
      summary: Performs user registration
      operationId: SignUpPost
      requestBody:
        description: Data to add new user to the social network
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: user1
                email:
                  type: string
                  example: user1@gmail.com
                password:
                  type: string
                  example: 123456789
              required:
                - username
                - email
                - password
      responses:
        '201':
          description: Successful operation
        default:
          $ref: '#/components/responses/code_error'
  /sign_in:
    post:
      tags:
        - sign
      summary: Performs user authentication
      operationId: SignInPost
      requestBody:
        description: Data to perform sign in to the social network
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: user1@gmail.com
                password:
                  type: string
                  example: 123456789
              required:
                - email
                - password
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  userResponse:
                    $ref: '#/components/schemas/user'
                  accessToken:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
                  refreshToken:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
                required:
                  - userResponse
                  - accessToken
                  - refreshToken
        default:
          $ref: '#/components/responses/code_error'
  /sign_refresh:
    post:
      tags:
        - sign
      summary: Exchanges a refresh token for an access token
      operationId: SignRefreshPost
      requestBody:
        description: Refresh token
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
              required:
                - refreshToken
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
                required:
                  - accessToken
        default:
          $ref: '#/components/responses/code_error'
  /me:
    get:
      tags:
        - me
      summary: Returns current user profile
      operationId: MeGet
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/user'
        default:
          $ref: '#/components/responses/code_error'
    put:
      tags:
        - me
      summary: Updates current user profile
      operationId: MePut
      security:
        - bearerAuth: []
      requestBody:
        description: Data to update the current user profile
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  example: 48690945-c3f9-45b4-9045-c6bcfb898f20
                username:
                  type: string
                  example: johndoe
                fullname:
                  type: string
                  example: John Doe
                status:
                  type: string
                  example: active
              required:
                - user_id
      responses:
        '204':
          description: Successful operation
        default:
          $ref: '#/components/responses/code_error'
    delete:
      tags:
        - me
      summary: Deletes current user profile
      operationId: MeDel
      security:
        - bearerAuth: []
      requestBody:
        description: Data to delete the current user profile
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  example: 48690945-c3f9-45b4-9045-c6bcfb898f20
              required:
                - user_id
      responses:
        '204':
          description: Successful operation
        default:
          $ref: '#/components/responses/code_error'
  /me/jokes:
    get:
      tags:
        - me
      summary: Returns jokes of the current user
      operationId: MeJokesGet
      parameters:
        - $ref: '#/components/parameters/select'
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/jokes'
        default:
          $ref: '#/components/responses/code_error'
    post:
      tags:
        - me
      summary: Creates a new joke for current user
      operationId: MeJokesPost
      security:
        - bearerAuth: []
      requestBody:
        description: Data to create a new joke for current user
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Joke title
                text:
                  type: string
                  example: Joke text
                explanation:
                  type: string
                  example: Joke explanation
              required:
                - user_id
                - title
                - text
      responses:
        '201':
          description: Joke created
        default:
          $ref: '#/components/responses/code_error'
    put:
      tags:
        - me
      summary: Updates joke for current user
      operationId: MeJokesPut
      security:
        - bearerAuth: []
      requestBody:
        description: Data to update the joke for current user
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                joke_id:
                  type: string
                  example: 48690945-c3f9-45b4-9045-c6bcfb898f20
                title:
                  type: string
                  example: Joke title
                text:
                  type: string
                  example: Joke text
                explanation:
                  type: string
                  example: Joke explanation
              required:
                - joke_id
      responses:
        '201':
          description: Joke updated
        default:
          $ref: '#/components/responses/code_error'
    delete:
      tags:
        - me
      summary: Deletes joke for current user
      operationId: MeJokesDel
      security:
        - bearerAuth: []
      requestBody:
        description: Data to delete the joke for current user
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                joke_id:
                  type: string
                  example: 48690945-c3f9-45b4-9045-c6bcfb898f20
              required:
                - joke_id
      responses:
        '204':
          description: Joke deleted
        default:
          $ref: '#/components/responses/code_error'
  /me/comments:
    post:
      tags:
        - me
      summary: Creates a comment for the current user and the current joke
      operationId: MeCommentsPost
      parameters:
        - $ref: '#/components/parameters/select'
      security:
        - bearerAuth: []
      requestBody:
        description: Create Comment Request
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                joke_id:
                  type: string
                  example: '1'
                text:
                  type: string
                  example: Text of comment 1
              required:
                - user_id
                - joke_id
                - text
      responses:
        '200':
          description: Comment created
        default:
          $ref: '#/components/responses/code_error'
    put:
      tags:
        - me
      summary: Updates a comment of the current user
      operationId: MeCommentsPut
      security:
        - bearerAuth: []
      requestBody:
        description: Update Comment Request
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                comment_id:
                  type: string
                  example: '1'
                text:
                  type: string
                  example: Text of comment 1
              required:
                - comment_id
      responses:
        '200':
          description: Comment updated
        default:
          $ref: '#/components/responses/code_error'
    delete:
      tags:
        - me
      summary: Deletes a comment of the current user
      operationId: MeCommentsDel
      security:
        - bearerAuth: []
      requestBody:
        description: Delete Comment Request
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                comment_id:
                  type: string
                  example: '1'
              required:
                - comment_id
      responses:
        '204':
          description: Comment deleted
        default:
          $ref: '#/components/responses/code_error'
  /me/likes:
    post:
      tags:
        - me
      summary: Creates a like for a joke for the current user
      operationId: MeLikesPost
      security:
        - bearerAuth: []
      requestBody:
        description: Data to create a like for a joke for the current user
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                joke_id:
                  type: string
                  example: 48690945-c3f9-45b4-9045-c6bcfb898f20
              required:
                - joke_id
      responses:
        '201':
          description: Like created
        default:
          $ref: '#/components/responses/code_error'
    delete:
      tags:
        - me
      summary: Deletes a like of the current user
      operationId: MeLikesDel
      security:
        - bearerAuth: []
      requestBody:
        description: Data to delete a like of the current user
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                joke_id:
                  type: string
                  example: 48690945-c3f9-45b4-9045-c6bcfb898f20
              required:
                - joke_id
      responses:
        '204':
          description: Like deleted
        default:
          $ref: '#/components/responses/code_error'
  /comments/{joke_id}:
    get:
      tags:
        - comments
      summary: Returns comments of the joke
      operationId: CommentsByJokeIDGet
      parameters:
        - name: joke_id
          in: path
          description: ID of the joke to get comments
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/select'
      responses:
        '200':
          $ref: '#/components/responses/comments'
        default:
          $ref: '#/components/responses/code_error'
  /likes/{joke_id}:
    get:
      tags:
        - likes
      summary: Counts likes of the joke
      operationId: LikesByJokeIDGet
      parameters:
        - name: joke_id
          in: path
          description: ID of the joke to count likes
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Count of likes
          content:
            application/json:
              schema:
                type: integer
                minimum: 0
        default:
          $ref: '#/components/responses/code_error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    code_error:
      type: object
      title: CodeError
      properties:
        code:
          type: string
          enum:
            - unauthenticated
            - invalid_argument
            - not_found
            - already_exists
            - permission_denied
            - internal
        message:
          type: string
      required:
        - code
        - message
    user:
      type: object
      properties:
        id:
          type: string
          example: 48690945-c3f9-45b4-9045-c6bcfb898f20
        username:
          type: string
          example: user1
        email:
          type: string
          example: user1@gmail.com
        fullname:
          type: string
          example: User 1
        status:
          type: string
          example: Just coding
        created_at:
          type: string
          format: date-time
          example: '2021-08-01T00:00:00Z'
      required:
        - id
        - username
        - email
        - created_at
    order:
      type: string
      name: order
      enum:
        - asc
        - desc
    joke:
      type: object
      properties:
        id:
          type: string
        userID:
          type: string
        title:
          type: string
          example: Joke 1
        text:
          type: string
          example: Text of joke 1
        explanation:
          type: string
          example: Explanation of joke 1
        createdAt:
          type: string
          format: date-time
          example: '2021-08-01T00:00:00Z'
      required:
        - id
        - userID
        - title
        - text
        - createdAt
    comment:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        joke_id:
          type: string
        text:
          type: string
          example: Text of comment 1
        createdAt:
          type: string
          format: date-time
          example: '2021-08-01T00:00:00Z'
      required:
        - id
        - user_id
        - joke_id
        - text
        - createdAt
  responses:
    code_error:
      description: General Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/code_error'
    user:
      description: User profile
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/user'
    jokes:
      description: Array of jokes
      content:
        application/json:
          schema:
            type: object
            properties:
              jokes:
                type: array
                items:
                  $ref: '#/components/schemas/joke'
            required:
              - jokes
    comments:
      description: Array of comments
      content:
        application/json:
          schema:
            type: object
            properties:
              comments:
                type: array
                items:
                  $ref: '#/components/schemas/comment'
            required:
              - comments
  parameters:
    select:
      name: select_params
      in: query
      description: fields to specify select parameters
      required: true
      schema:
        type: object
        title: select_params
        properties:
          order_by:
            type: string
            default: created_at
          order:
            $ref: '#/components/schemas/order'
          limit:
            type: integer
            minimum: 0
          offset:
            type: integer
            minimum: 0
        required:
          - limit
          - offset